+++ service/frauddetector/errors.go
@@ -8,6 +8,23 @@ import (
 
 const (
 
+	// ErrCodeConflictException for service response error code
+	// "ConflictException".
+	//
+	// An exception indicating there was a conflict during a delete operation. The
+	// following delete operations can cause a conflict exception:
+	//
+	//    * DeleteDetector: A conflict exception will occur if the detector has
+	//    associated Rules or DetectorVersions. You can only delete a detector if
+	//    it has no Rules or DetectorVersions.
+	//
+	//    * DeleteDetectorVersion: A conflict exception will occur if the DetectorVersion
+	//    status is ACTIVE.
+	//
+	//    * DeleteRuleVersion: A conflict exception will occur if the RuleVersion
+	//    is in use by an associated ACTIVE or INACTIVE DetectorVersion.
+	ErrCodeConflictException = "ConflictException"
+
 	// ErrCodeInternalServerException for service response error code
 	// "InternalServerException".
 	//
@@ -34,6 +51,7 @@ const (
 )
 
 var exceptionFromCode = map[string]func(protocol.ResponseMetadata) error{
+	"ConflictException":         newErrorConflictException,
 	"InternalServerException":   newErrorInternalServerException,
 	"ResourceNotFoundException": newErrorResourceNotFoundException,
 	"ThrottlingException":       newErrorThrottlingException,